requirements.txt
Django==4.2.13
django-filter==24.2
djangorestframework==3.15.1
Markdown==3.6
psycopg2-binary==2.9.9
django-dotenv==1.4.2
django-filter==24.2